[MAT-278] Split dispatch from runners
This is so we can have a stab at generating runners first.
